Revert the normalize() float(_amp) change; re-ledger the AA tests that depended on it
The earlier `Potential.normalize -> float(_amp)` broke the deliberate "amp stays
on-backend (differentiable)" contract asserted by the pillar-2 backend tests
(test_backend_potential_convenience.py::test_normalize_torch/jax_backend,
test_backend_ellipsoidal.py::test_construct_normalize_amp_is_float64, from #989).

Revert to `self._amp *= norm / abs(Rforce(1,0))` so normalize() under a forced
backend keeps producing a float64 BACKEND-array amp. Re-ledger the AA tests that
flipped only because of the float() amp (in-test `Pot(normalize=...)` evaluated
through the numpy island: Spherical otherIsochrone/basic_freqs/linear_angles,
Staeckel/grid otherIsochrone, etc.): 18 torch + 5 jax.

Net AA un-ledger this PR: 91 torch + 70 jax (was 106/75). torch test_actionAngle
0 failed; jax 0 failed; the 6 pillar-2 normalize tests pass again.

The proper way to flip those back (a forced-backend-constructed potential should
evaluate consistently under a forced-numpy island) is eval-time coercion of the
stored params to the active namespace -- a broader potential-funnel change left
to a follow-up, not forced into this PR.

+++ galpy/potential/Potential.py
@@ -993,13 +993,8 @@ class Potential(Force):
 
         """
         # abs() (via __abs__) is backend-agnostic and byte-identical to the old
-        # numpy.fabs on the numpy scalar Rforce returns. float() keeps the
-        # amplitude a backend-agnostic scalar: under a forced jax/torch backend
-        # Rforce(1,0) returns a backend array, and storing that in self._amp
-        # would pin the whole potential to that backend (every later eval would
-        # return backend arrays even on a numpy/scalar query). The normalization
-        # is a single scalar at (R,z)=(1,0), so float() is exact/byte-identical.
-        self._amp *= float(norm / abs(self.Rforce(1.0, 0.0, use_physical=False)))
+        # numpy.fabs on the numpy scalar Rforce returns.
+        self._amp *= norm / abs(self.Rforce(1.0, 0.0, use_physical=False))
 
     def toPlanar(self):
         """
